refactor(rpc): move server initialization into rpc module
